+++ test_with_arrow/src/tests/arrays.rs
@@ -1504,7 +1504,7 @@ mod dictionary {
                 "a", "a", "b", "c", "c",
             ])?,
             Array::Dictionary(DictionaryArray {
-                indices: Box::new(Array::Int8(PrimitiveArray {
+                keys: Box::new(Array::Int8(PrimitiveArray {
                     validity: None,
                     values: vec![0, 0, 1, 2, 2],
                 })),
@@ -1528,7 +1528,7 @@ mod dictionary {
                 Some("c"),
             ])?,
             Array::Dictionary(DictionaryArray {
-                indices: Box::new(Array::Int8(PrimitiveArray {
+                keys: Box::new(Array::Int8(PrimitiveArray {
                     validity: Some(vec![0b_11001]),
                     values: vec![0, 0, 0, 1, 1],
                 })),
@@ -1548,7 +1548,7 @@ mod dictionary {
                 "a", "a", "b", "c", "c",
             ])?,
             Array::Dictionary(DictionaryArray {
-                indices: Box::new(Array::Int16(PrimitiveArray {
+                keys: Box::new(Array::Int16(PrimitiveArray {
                     validity: None,
                     values: vec![0, 0, 1, 2, 2],
                 })),
@@ -1572,7 +1572,7 @@ mod dictionary {
                 Some("c"),
             ])?,
             Array::Dictionary(DictionaryArray {
-                indices: Box::new(Array::Int16(PrimitiveArray {
+                keys: Box::new(Array::Int16(PrimitiveArray {
                     validity: Some(vec![0b_11001]),
                     values: vec![0, 0, 0, 1, 1],
                 })),
@@ -1592,7 +1592,7 @@ mod dictionary {
                 "a", "a", "b", "c", "c",
             ])?,
             Array::Dictionary(DictionaryArray {
-                indices: Box::new(Array::Int32(PrimitiveArray {
+                keys: Box::new(Array::Int32(PrimitiveArray {
                     validity: None,
                     values: vec![0, 0, 1, 2, 2],
                 })),
@@ -1616,7 +1616,7 @@ mod dictionary {
                 Some("c"),
             ])?,
             Array::Dictionary(DictionaryArray {
-                indices: Box::new(Array::Int32(PrimitiveArray {
+                keys: Box::new(Array::Int32(PrimitiveArray {
                     validity: Some(vec![0b_11001]),
                     values: vec![0, 0, 0, 1, 1],
                 })),
@@ -1636,7 +1636,7 @@ mod dictionary {
                 "a", "a", "b", "c", "c",
             ])?,
             Array::Dictionary(DictionaryArray {
-                indices: Box::new(Array::Int64(PrimitiveArray {
+                keys: Box::new(Array::Int64(PrimitiveArray {
                     validity: None,
                     values: vec![0, 0, 1, 2, 2],
                 })),
@@ -1660,7 +1660,7 @@ mod dictionary {
                 Some("c"),
             ])?,
             Array::Dictionary(DictionaryArray {
-                indices: Box::new(Array::Int64(PrimitiveArray {
+                keys: Box::new(Array::Int64(PrimitiveArray {
                     validity: Some(vec![0b_11001]),
                     values: vec![0, 0, 0, 1, 1],
                 })),
@@ -1680,7 +1680,7 @@ mod dictionary {
                 "a", "a", "b", "c", "c",
             ])?,
             Array::Dictionary(DictionaryArray {
-                indices: Box::new(Array::UInt8(PrimitiveArray {
+                keys: Box::new(Array::UInt8(PrimitiveArray {
                     validity: None,
                     values: vec![0, 0, 1, 2, 2],
                 })),
@@ -1704,7 +1704,7 @@ mod dictionary {
                 Some("c"),
             ])?,
             Array::Dictionary(DictionaryArray {
-                indices: Box::new(Array::UInt8(PrimitiveArray {
+                keys: Box::new(Array::UInt8(PrimitiveArray {
                     validity: Some(vec![0b_11001]),
                     values: vec![0, 0, 0, 1, 1],
                 })),
@@ -1724,7 +1724,7 @@ mod dictionary {
                 "a", "a", "b", "c", "c",
             ])?,
             Array::Dictionary(DictionaryArray {
-                indices: Box::new(Array::UInt16(PrimitiveArray {
+                keys: Box::new(Array::UInt16(PrimitiveArray {
                     validity: None,
                     values: vec![0, 0, 1, 2, 2],
                 })),
@@ -1748,7 +1748,7 @@ mod dictionary {
                 Some("c"),
             ])?,
             Array::Dictionary(DictionaryArray {
-                indices: Box::new(Array::UInt16(PrimitiveArray {
+                keys: Box::new(Array::UInt16(PrimitiveArray {
                     validity: Some(vec![0b_11001]),
                     values: vec![0, 0, 0, 1, 1],
                 })),
@@ -1768,7 +1768,7 @@ mod dictionary {
                 "a", "a", "b", "c", "c",
             ])?,
             Array::Dictionary(DictionaryArray {
-                indices: Box::new(Array::UInt32(PrimitiveArray {
+                keys: Box::new(Array::UInt32(PrimitiveArray {
                     validity: None,
                     values: vec![0, 0, 1, 2, 2],
                 })),
@@ -1792,7 +1792,7 @@ mod dictionary {
                 Some("c"),
             ])?,
             Array::Dictionary(DictionaryArray {
-                indices: Box::new(Array::UInt32(PrimitiveArray {
+                keys: Box::new(Array::UInt32(PrimitiveArray {
                     validity: Some(vec![0b_11001]),
                     values: vec![0, 0, 0, 1, 1],
                 })),
@@ -1812,7 +1812,7 @@ mod dictionary {
                 "a", "a", "b", "c", "c",
             ])?,
             Array::Dictionary(DictionaryArray {
-                indices: Box::new(Array::UInt64(PrimitiveArray {
+                keys: Box::new(Array::UInt64(PrimitiveArray {
                     validity: None,
                     values: vec![0, 0, 1, 2, 2],
                 })),
@@ -1836,7 +1836,7 @@ mod dictionary {
                 Some("c"),
             ])?,
             Array::Dictionary(DictionaryArray {
-                indices: Box::new(Array::UInt64(PrimitiveArray {
+                keys: Box::new(Array::UInt64(PrimitiveArray {
                     validity: Some(vec![0b_11001]),
                     values: vec![0, 0, 0, 1, 1],
                 })),
@@ -1844,6 +1844,66 @@ mod dictionary {
                     validity: None,
                     offsets: vec![0, 1, 2],
                     data: b"ac".to_vec(),
+                })),
+            }),
+        )
+    }
+}
+
+mod run_end_encoded {
+    use marrow::{
+        array::{BytesArray, RunEndEncodedArray},
+        datatypes::RunEndEncodedMeta,
+    };
+
+    use super::*;
+
+    #[test]
+    fn not_nullable() -> PanicOnError<()> {
+        assert_arrays_eq(
+            Arc::new(
+                arrow_array::RunArray::<arrow_array::types::Int32Type>::from_iter([
+                    "hello", "hello", "world", "foo",
+                ]),
+            ) as arrow_array::ArrayRef,
+            Array::RunEndEncoded(RunEndEncodedArray {
+                meta: RunEndEncodedMeta::default(),
+                run_ends: Box::new(Array::Int32(PrimitiveArray {
+                    validity: None,
+                    values: vec![2, 3, 4],
+                })),
+                values: Box::new(Array::Utf8(BytesArray {
+                    validity: None,
+                    offsets: vec![0, 5, 10, 13],
+                    data: b"helloworldfoo".to_vec(),
+                })),
+            }),
+        )
+    }
+
+    #[test]
+    fn nullable() -> PanicOnError<()> {
+        assert_arrays_eq(
+            Arc::new(
+                arrow_array::RunArray::<arrow_array::types::Int32Type>::from_iter([
+                    Some("hello"),
+                    Some("hello"),
+                    None,
+                    None,
+                    Some("world"),
+                    Some("foo"),
+                ]),
+            ) as arrow_array::ArrayRef,
+            Array::RunEndEncoded(RunEndEncodedArray {
+                meta: RunEndEncodedMeta::default(),
+                run_ends: Box::new(Array::Int32(PrimitiveArray {
+                    validity: None,
+                    values: vec![2, 4, 5, 6],
+                })),
+                values: Box::new(Array::Utf8(BytesArray {
+                    validity: Some(vec![0b1101]),
+                    offsets: vec![0, 5, 5, 10, 13],
+                    data: b"helloworldfoo".to_vec(),
                 })),
             }),
         )