+++ Changes.md
@@ -4,12 +4,18 @@
 
 - Rework map arrays to use explicit keys and values array to simplify interaction the underlying
   arrays
-- Rework `StructArray` and `DenseUnionArray`: place metadata in front of arrays in
-  `StructArray::fields`, `DenseUnionArray::fields`
+- Implement sparse unions, rename `DenseUnion` to `Union` and change offsets to be `Option<Vec<i32>>`
+- Implement interval arrays and the `Interval` data type
+- Implement run encoded array
+- Rename `Dictionary::indices` to `Dictionary::keys`
+- Rework `StructArray` and `UnionArray`: place metadata in front of arrays in `StructArray::fields`,
+  `UnionArray::fields`
 - Add `MarrowError::new` and `MarrowError::with_cause`
 - Add `as_view` for `Array` and the array structs
 - Implement `PartialEq` for `Array` and `View`, and `FieldMeta`
 - Implement `Default` for `Field` and `FieldMeta`
+- Remove the sorted flag from the dictionary `DataType` it is not supported by `arrow`
+- Add `Array::data_type()` and  `View::data_type()`
 
 ## 0.1.0